Refactor and improve config.InterfacesFile
Split the file:line tracking code is to a new class: FilePosition.  This
allows those objects to be cloned so errors occurring after intial
parsing can still be attributed to an appropriate location in the
configuration file(s).

Also, make config.InterfacesFile default to automatically closing the
input files when it's done reading them.  This can be controlled by the
paramter "autoclose=(True|False)".

Finally, add lots of new docstrings.

Signed-off-by: Kyle Moffett <kyle@moffetthome.net>