New module for executing helper scripts in a clean environment
To ensure that helper scripts are idempotent and reproducible, their
execution environment needs to be carefully controlled.  That means the
working directory, various environment variables, etc.

This new "ifupdown.script" module contains several classes for creating
this clean environment for scripts.

It currently isn't used anywhere yet, but I've verified that it has the
expected behavior with ad-hoc tests (real unit-tests are still TODO), so
it should get checked in.

Signed-off-by: Kyle Moffett <kyle@moffetthome.net>